# build/build/swagger.json
{
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "UserType": {
                "properties": {
                    "id": {
                        "type": "number",
                        "format": "double"
                    },
                    "name": {
                        "type": "string"
                    },
                    "cellphone": {
                        "type": "string"
                    }
                },
                "required": [
                    "id"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "Pick_UserType.id-or-name-or-cellphone_": {
                "properties": {
                    "id": {
                        "type": "number",
                        "format": "double"
                    },
                    "name": {
                        "type": "string"
                    },
                    "cellphone": {
                        "type": "string"
                    }
                },
                "required": [
                    "id"
                ],
                "type": "object",
                "description": "From T, pick a set of properties whose keys are in the union K"
            },
            "UserCreationParams": {
                "$ref": "#/components/schemas/Pick_UserType.id-or-name-or-cellphone_"
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "rest-api-ts-tsoa",
        "version": "1.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
        "/users/{userId}": {
            "get": {
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserType"
                                }
                            }
                        }
                    }
                },
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/users": {
            "post": {
                "operationId": "CreateUser",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreationParams"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}